# simpeg-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Permission:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  domain.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  http.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  http.RoleStoreRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8000
info:
  contact:
    name: SIMPEG UML
  description: Backend API Sistem Informasi Manajemen Kepegawaian Universitas Muhammadiyah
    Lampung (Gin + GORM).
  termsOfService: http://simpeg.uml.ac.id
  title: SIMPEG UML API
  version: "1.0"
paths:
...
  /roles:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Role'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar role
      tags:
      - Roles & Permissions
    post:
      consumes:
      - application/json
      parameters:
      - description: Data role (name, description, permissions)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.RoleStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat role baru
      tags:
      - Roles & Permissions
  /roles/{name}:
    delete:
      description: Role `admin` tidak dapat dihapus.
      parameters:
      - description: Nama role
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.MessageResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus role
      tags:
      - Roles & Permissions
    get:
      parameters:
      - description: Nama role
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Role'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detail role berdasarkan nama
      tags:
      - Roles & Permissions
    put:
      consumes:
      - application/json
      parameters:
      - description: Nama role
        in: path
        name: name
        required: true
        type: string
      - description: Data role baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/http.RoleStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Role'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ubah role
      tags:
      - Roles & Permissions
  /roles/permissions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Permission'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Daftar semua permission
      tags:
      - Roles & Permissions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"